# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  description: API do app MoveUp. Erros em application/problem+json (RFC 9457) com
    `code` estável.
  title: MoveUp API
  version: v1
servers:
- url: /
security:
- bearerAuth: []
...
paths:
...
  /v1/anamnesis/template:
    get:
      operationId: getAnamnesisTemplate
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnamnesisTemplate"
          description: Modelo do sistema com PAR-Q
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`resource-not-found` (vínculo de outro profissional, versã\
            o inexistente)"
      summary: Perguntas da anamnese (modelo atual)
      tags:
      - anamnesis
...
  /v1/clients/{linkId}/anamnesis:
    get:
      description: Abrir fica registrado na trilha de auditoria (view_anamnesis).
      operationId: getClientAnamnesis
      parameters:
      - in: path
        name: linkId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientAnamnesis"
          description: Mais recente e lista de versões
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`resource-not-found` (vínculo de outro profissional, versã\
            o inexistente)"
      summary: Anamnese do aluno (mais recente e versões)
      tags:
      - anamnesis
  /v1/clients/{linkId}/anamnesis/review:
    post:
      description: Revisada fica imutável; revisar de novo cria uma versão nova.
      operationId: reviewClientAnamnesis
      parameters:
      - in: path
        name: linkId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnamnesisReview"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anamnesis"
          description: Versão revisada
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`resource-not-found` (vínculo de outro profissional, versã\
            o inexistente)"
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`answer-invalid`, `clearance-invalid`, `clearance-date-required`"
      summary: Revisa a anamnese (complemento e liberação médica)
      tags:
      - anamnesis
  /v1/clients/{linkId}/anamnesis/versions/{versionNumber}:
    get:
      operationId: getClientAnamnesisVersion
      parameters:
      - in: path
        name: linkId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: versionNumber
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anamnesis"
          description: A versão
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`resource-not-found` (vínculo de outro profissional, versã\
            o inexistente)"
      summary: Uma versão da anamnese
      tags:
      - anamnesis
...
  /v1/clients/{linkId}/restrictions:
    get:
      operationId: listRestrictions
      parameters:
      - in: path
        name: linkId
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: includeResolved
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HealthRestriction"
          description: Restrições
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`resource-not-found` (vínculo de outro profissional, versã\
            o inexistente)"
      summary: Restrições do aluno (ativas primeiro)
      tags:
      - anamnesis
    post:
      operationId: createRestriction
      parameters:
      - in: path
        name: linkId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HealthRestrictionInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthRestriction"
          description: Criada
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`resource-not-found` (vínculo de outro profissional, versã\
            o inexistente)"
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`restriction-kind-invalid`, `restriction-region-invalid`,\
            \ `restriction-description-invalid`, `restriction-severity-invalid`"
      summary: Nova restrição
      tags:
      - anamnesis
  /v1/clients/{linkId}/restrictions/{restrictionId}:
    delete:
      operationId: deleteRestriction
      parameters:
      - in: path
        name: linkId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: restrictionId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Excluída
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`resource-not-found` (vínculo de outro profissional, versã\
            o inexistente)"
      summary: Exclui uma restrição lançada errado
      tags:
      - anamnesis
    put:
      operationId: updateRestriction
      parameters:
      - in: path
        name: linkId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: restrictionId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HealthRestrictionInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthRestriction"
          description: Alterada
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`resource-not-found` (vínculo de outro profissional, versã\
            o inexistente)"
      summary: Altera ou resolve uma restrição
      tags:
      - anamnesis
...
  /v1/me/anamnesis:
    get:
      operationId: getMyAnamnesis
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyAnamnesis"
          description: "Sem anamnese, o campo vem ausente"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`resource-not-found` (vínculo de outro profissional, versã\
            o inexistente)"
      summary: A anamnese mais recente do aluno
      tags:
      - anamnesis
    put:
      description: "Antes da revisão do personal substitui a versão atual; depois,\
        \ cria uma versão nova."
      operationId: submitMyAnamnesis
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnamnesisSubmit"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anamnesis"
          description: Versão gravada
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`resource-not-found` (vínculo de outro profissional, versã\
            o inexistente)"
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`answer-invalid`, `answer-required`, `answer-unknown`, `anamnesis-without-link`"
      summary: Envia a anamnese
      tags:
      - anamnesis
...
components:
  schemas:
...
    Anamnesis:
      type: object
      properties:
        answers:
          type: object
          additionalProperties: {}
          description: "Código da pergunta → resposta (texto, sim/não, número ou lista)"
        clearance:
          type: string
          enum:
          - not_required
          - pending
          - cleared
        clearanceDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        parqPositive:
          type: boolean
        reviewed:
          type: boolean
        reviewedAt:
          type: string
          format: date-time
        versionNumber:
          type: integer
          format: int32
      required:
      - answers
      - clearance
      - createdAt
      - parqPositive
      - reviewed
      - versionNumber
    AnamnesisOption:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
      required:
      - label
      - value
    AnamnesisQuestion:
      type: object
      properties:
        code:
          type: string
        label:
          type: string
        max:
          type: integer
          format: int32
        min:
          type: integer
          format: int32
        options:
          type: array
          items:
            $ref: "#/components/schemas/AnamnesisOption"
        required:
          type: boolean
        section:
          type: string
          enum:
          - goal
          - routine
          - parq
          - health
        type:
          type: string
          enum:
          - single
          - multi
          - yes_no
          - integer
          - text
      required:
      - code
      - label
      - options
      - required
      - section
      - type
    AnamnesisReview:
      type: object
      properties:
        answers:
          type: object
          additionalProperties: {}
          description: Respostas com o complemento do personal
        clearance:
          type: string
          enum:
          - not_required
          - pending
          - cleared
          minLength: 1
        clearanceDate:
          type: string
          format: date
          description: Obrigatória quando clearance = cleared
      required:
      - answers
      - clearance
    AnamnesisSubmit:
      type: object
      properties:
        answers:
          type: object
          additionalProperties: {}
      required:
      - answers
    AnamnesisTemplate:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/AnamnesisQuestion"
        version:
          type: integer
          format: int32
      required:
      - questions
      - version
    AnamnesisVersion:
      type: object
      properties:
        clearance:
          type: string
          enum:
          - not_required
          - pending
          - cleared
        createdAt:
          type: string
          format: date-time
        parqPositive:
          type: boolean
        reviewedAt:
          type: string
          format: date-time
        versionNumber:
          type: integer
          format: int32
      required:
      - clearance
      - createdAt
      - parqPositive
      - versionNumber
...
    ClientAnamnesis:
      type: object
      properties:
        latest:
          $ref: "#/components/schemas/Anamnesis"
          description: Ausente se o aluno ainda não enviou
        versions:
          type: array
          items:
            $ref: "#/components/schemas/AnamnesisVersion"
      required:
      - versions
...
    FieldProblem:
      type: object
      properties:
        code:
          type: string
          example: min
        field:
          type: string
          example: reps.max
        message:
          type: string
      required:
      - code
      - field
      - message
...
    HealthRestriction:
      type: object
      properties:
        bodyRegion:
          type: string
          description: Mesma lista de regiões do relato de dor
        description:
          type: string
        fromAnamnesis:
          type: boolean
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
          - injury
          - surgery
          - pain
          - condition
        resolvedOn:
          type: string
          format: date
          description: Ausente = ativa
        severity:
          type: integer
          format: int32
          description: "1 leve, 2 moderada, 3 grave"
      required:
      - description
      - fromAnamnesis
      - id
      - kind
    HealthRestrictionInput:
      type: object
      properties:
        bodyRegion:
          type: string
        description:
          type: string
          maxLength: 500
          minLength: 0
        kind:
          type: string
          enum:
          - injury
          - surgery
          - pain
          - condition
          minLength: 1
        resolvedOn:
          type: string
          format: date
        severity:
          type: integer
          format: int32
      required:
      - description
      - kind
...
    MyAnamnesis:
      type: object
      properties:
        anamnesis:
          $ref: "#/components/schemas/Anamnesis"
          description: Ausente antes do primeiro envio
...
    Problem:
      type: object
      description: Erro no formato RFC 9457 (application/problem+json)
      properties:
        code:
          type: string
          description: Código estável (kebab-case) que o app usa para escolher a mensagem
          example: resource-not-found
        detail:
          type: string
          example: Recurso não encontrado.
        errors:
          type: array
          description: Só em erros de validação (400)
          items:
            $ref: "#/components/schemas/FieldProblem"
        instance:
          type: string
          example: /v1/me
        status:
          type: integer
          format: int32
          example: 404
        title:
          type: string
          example: Recurso não encontrado
        traceId:
          type: string
          example: 4bf92f3577b34da6a3ce929d0e0e4736
        type:
          type: string
          example: https://api.moveup.com.br/problems/resource-not-found
      required:
      - code
      - detail
      - status
      - title
      - traceId
      - type
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: ID token do Firebase Auth
      scheme: bearer
      type: http